requirement.txt
pandas==1.5.0
psycopg2-binary==2.9.3
tqdm==4.64.1
